Jornada Laboral

+++ Software2G/src/com/software2g/portal/facade/IGestionFacadePortal.java
@@ -72,7 +72,7 @@ public interface IGestionFacadePortal {
 	
 	//--------------------------------------------------------------------------
 	//Metodos Entidad Persona
-	public Persona findPersonaById(java.lang.Integer id) throws Exception;
+	public Persona findPersonaById(java.lang.Long id) throws Exception;
 	public List<Persona> findAllPersonas() throws Exception;
 	public void persistPersona(Persona persona) throws Exception;
 	public void removePersona(Persona persona) throws Exception;
